exceptions catching

+++ translator.py
@@ -5,9 +5,12 @@ import time
 import pandas as pd
 from openai import OpenAI
 from tqdm import tqdm
+from tenacity import retry, stop_after_attempt, wait_fixed, RetryError
+import traceback
 
 client = None
 
+@retry(stop=stop_after_attempt(5), wait=wait_fixed(10))
 def translate_text(model_name,text):  # 用来调用API翻译传入的文本
     try:
         response = client.chat.completions.create(
@@ -20,10 +23,12 @@ def translate_text(model_name,text):  # 用来调用API翻译传入的文本
             temperature=0.1,
             max_tokens=16384,
         )
-        return str(response.choices[0].message.content).replace(",","，").replace("\n"," ")
+        return str(response.choices[0].message.content).replace("\n"," ")
     except Exception as e:
-        print(f"Error: {e}")
-        return None
+        print('Exception occurred. Retrying in 10s...', file=sys.stderr)
+        print('Details:', file=sys.stderr)
+        traceback.print_exc()
+        raise e
 
 def usage():
     print("Usage: python translator.py [OPTION] <INPUT_FILE>")
@@ -37,7 +42,7 @@ def main():
     try:
         opts, args = getopt.getopt(sys.argv[1:], "ho:t:u:m:", ["help","output=","token=","url=","model="])
     except getopt.GetoptError as err:
-        print(err)
+        print(err,file=sys.stderr)
         usage()
         sys.exit(2)
     output = "result.csv"
@@ -69,24 +74,46 @@ def main():
         base_url=url,
         api_key=hf_token
     )
+    tqdm.write('Loading data...')
     # 读取数据
     df = pd.read_csv(inp)
-    rows = []
-    for index, row in df.iterrows():
-        rows.append(row)
-    tqdm.write(f'There are {len(rows)} rows in total.')
+    rows = len(df)
+    tqdm.write(f'There are {rows} rows in total.')
+    titles_out = []
+    if os.path.exists(output):
+        df_out = pd.read_csv(output)
+        for index, row in df_out.iterrows():
+            titles_out.append(row['title'])
+    else:
+        with open(output, "w", encoding="utf-8") as f:
+            f.write('title,authors,abstract,date,paper_url,score,title_cn,abstract_cn\n')
     # 遍历每一行，翻译摘要
-    with open(output, "w", encoding="utf-8") as f:
-        for index, row in enumerate(tqdm(rows, position=0, file=sys.stdout, desc="Translation Process:")):
-            abstract = row['abstract']
-            tqdm.write(f"Translating paper #{index+1}...")
-            cn_abstract = translate_text(model_name,abstract)
-            if cn_abstract is None:
-                pass #TODO:异常处理，明天再写
-            # 将结果写入新文件
-            f.write(f"{index+1},{cn_abstract}\n")
-            time.sleep(1)  # 避免请求过于频繁
-    tqdm.write('Translation completed.')
+    with open(output, "a", encoding="utf-8") as f:
+        try:
+            for index, row in tqdm(df.iterrows(),total=rows, position=0, file=sys.stdout, desc="Translation Process"):
+                title = row['title']
+                if title in titles_out:
+                    tqdm.write(f'Skipping existing paper #{index+1} (Title:\"{title}\")')
+                    continue
+                abstract = row['abstract']
+                tqdm.write(f"Translating paper #{index+1}...")
+                cn_title = translate_text(model_name, title)
+                cn_abstract = translate_text(model_name,abstract)
+                # 将结果写入新文件
+                f.write(f"\"{str(title).replace("\"","\"\"")}\","
+                        f"\"{str(row['authors']).replace("\"","\"\"")}\","
+                        f"\"{str(abstract).replace("\"","\"\"")}\","
+                        f"\"{str(row['date']).replace("\"","\"\"")}\","
+                        f"\"{str(row['paper_url']).replace("\"","\"\"")}\","
+                        f"\"{str(row['score']).replace("\"","\"\"")}\","
+                        f"\"{str(cn_title).replace("\"","\"\"")}\","
+                        f"\"{str(cn_abstract).replace("\"","\"\"")}\"\n")
+                f.flush() # 及时更新文件内容
+                time.sleep(1)  # 避免请求过于频繁
+            tqdm.write('Translation completed.')
+        except RetryError:
+            print('Retried 5 times. Translation failed.',file=sys.stderr)
+            exit(-1) # 直接退出程序，反正有断点续传
 
 if __name__ == "__main__":
     main()
